Heat juno-2 milestone (2014.2.b2)

+++ heat/engine/resources/nova_utils.py
@@ -16,17 +16,16 @@ import email
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 import json
+from novaclient import exceptions as nova_exceptions
 import os
 import pkgutil
 import string
 
-from novaclient import exceptions as nova_exceptions
 from oslo.config import cfg
 import six
 from six.moves.urllib import parse as urlparse
 
 from heat.common import exception
-from heat.engine import clients
 from heat.engine import scheduler
 from heat.openstack.common.gettextutils import _
 from heat.openstack.common import log as logging
@@ -52,15 +51,16 @@ def refresh_server(server):
     '''
     try:
         server.get()
-    except clients.novaclient.exceptions.OverLimit as exc:
+    except nova_exceptions.OverLimit as exc:
         msg = _("Server %(name)s (%(id)s) received an OverLimit "
                 "response during server.get(): %(exception)s")
         LOG.warning(msg % {'name': server.name,
                            'id': server.id,
                            'exception': exc})
-    except clients.novaclient.exceptions.ClientException as exc:
-        if ((getattr(exc, 'http_status', getattr(exc, 'code', None)) in
-             (500, 503))):
+    except nova_exceptions.ClientException as exc:
+        http_status = (getattr(exc, 'http_status', None) or
+                       getattr(exc, 'code', None))
+        if http_status in (500, 503):
             msg = _('Server "%(name)s" (%(id)s) received the following '
                     'exception during server.get(): %(exception)s')
             LOG.warning(msg % {'name': server.name,
@@ -185,7 +185,7 @@ echo -e '%s\tALL=(ALL)\tNOPASSWD: ALL' >> /etc/sudoers
         userdata_parts = None
         try:
             userdata_parts = email.message_from_string(userdata)
-        except:
+        except Exception:
             pass
         if userdata_parts and userdata_parts.is_multipart():
             for part in userdata_parts.get_payload():
@@ -238,18 +238,37 @@ echo -e '%s\tALL=(ALL)\tNOPASSWD: ALL' >> /etc/sudoers
 
 def delete_server(server):
     '''
-    Return a co-routine that deletes the server and waits for it to
+    A co-routine that deletes the server and waits for it to
     disappear from Nova.
     '''
-    server.delete()
+    if not server:
+        return
+    try:
+        server.delete()
+    except nova_exceptions.NotFound:
+        return
 
     while True:
         yield
 
         try:
             refresh_server(server)
-        except clients.novaclient.exceptions.NotFound:
+        except nova_exceptions.NotFound:
             break
+        else:
+            # Some clouds append extra (STATUS) strings to the status
+            short_server_status = server.status.split('(')[0]
+            if short_server_status == "DELETED":
+                break
+            if short_server_status == "ERROR":
+                fault = getattr(server, 'fault', {})
+                message = fault.get('message', 'Unknown')
+                code = fault.get('code')
+                errmsg = (_("Server %(name)s delete failed: (%(code)s) "
+                            "%(message)s"))
+                raise exception.Error(errmsg % {"name": server.name,
+                                                "code": code,
+                                                "message": message})
 
 
 @scheduler.wrappertask
@@ -330,7 +349,7 @@ def server_to_ipaddress(client, server):
     '''
     try:
         server = client.servers.get(server)
-    except clients.novaclient.exceptions.NotFound as ex:
+    except nova_exceptions.NotFound as ex:
         LOG.warn(_('Instance (%(server)s) not found: %(ex)s')
                  % {'server': server, 'ex': ex})
     else:
